requirements.txt
pandas
numpy
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-dotenv
